requirements.txt
aiohttp
websockets
pandas
numpy
python-telegram-bot
python-dotenv